requirements.txt
pyTelegramBotAPI
urllib
opencv
numpy
pillow
torch
torchvision
albumentations